simplified page to make it more self containing
